+++ projects/rocprofiler-systems/source/lib/backends/amd_smi/device_backend.hpp
@@ -4,12 +4,15 @@
 #pragma once
 
 #include "backends/amd_smi/gpu_types.hpp"
+#include "backends/amd_smi/nic_types.hpp"
 #include "backends/amd_smi/sdma_feature.hpp"
 
 #include <algorithm>
 #include <cstdint>
+#include <memory>
 #include <stdexcept>
 #include <string>
+#include <string_view>
 #include <vector>
 
 #include <amd_smi/amdsmi.h>
@@ -24,10 +27,21 @@ using gpu::MAX_NUM_XGMI_LINKS;
 using gpu::metrics;
 using gpu::populate_if_supported;
 
-class gpu_backend
+/**
+ * @brief AMD SMI backend that owns a single processor handle.
+ *
+ * One instance wraps one @c amdsmi_processor_handle and encapsulates all AMD SMI
+ * details (handle, status codes, raw structs). The same type serves both GPU and
+ * AI NIC devices: the GPU collector uses the @c get_gpu_* methods, the NIC
+ * collector uses the @c get_nic_* methods, and each is checked against its own
+ * contract in the PMC layer.
+ *
+ * Methods throw std::runtime_error on AMD SMI failures.
+ */
+class device_backend
 {
 public:
-    explicit gpu_backend(amdsmi_processor_handle handle) noexcept
+    explicit device_backend(amdsmi_processor_handle handle) noexcept
     : m_handle{ handle }
     {}
 
@@ -100,8 +114,89 @@ public:
 #endif
     }
 
+#if defined(ROCPROFSYS_BUILD_AINIC) && ROCPROFSYS_BUILD_AINIC == 1
+    /**
+     * @brief Get NIC ASIC information.
+     * @return ASIC info with vendor and product names.
+     * @throws std::runtime_error If AMD SMI query fails.
+     */
+    [[nodiscard]] nic::asic_info get_nic_asic_info() const
+    {
+        amdsmi_nic_asic_info_t raw{};
+        check(amdsmi_get_nic_asic_info(m_handle, &raw), "get_nic_asic_info");
+        return { raw.product_name, raw.vendor_name };
+    }
+
+    /**
+     * @brief Get NIC port information.
+     * @return Port info with device name (empty if no ports).
+     * @throws std::runtime_error If AMD SMI query fails.
+     */
+    [[nodiscard]] nic::port_info get_nic_port_info() const
+    {
+        amdsmi_nic_port_info_t raw{};
+        check(amdsmi_get_nic_port_info(m_handle, &raw), "get_nic_port_info");
+        if(raw.num_ports == 0)
+        {
+            return {};
+        }
+        return { raw.ports[0].netdev };
+    }
+
+    /**
+     * @brief Get NIC RDMA device information.
+     * @return RDMA info with port count (zero if no RDMA devices).
+     * @throws std::runtime_error If AMD SMI query fails.
+     */
+    [[nodiscard]] nic::rdma_info get_nic_rdma_info() const
+    {
+        // Heap-allocate: amdsmi_nic_rdma_devices_info_t is ~558 KiB (above the
+        // -Wstack-usage=524288 ceiling).
+        auto raw = std::make_unique<amdsmi_nic_rdma_devices_info_t>();
+        check(amdsmi_get_nic_rdma_dev_info(m_handle, raw.get()), "get_nic_rdma_dev_info");
+        if(raw->num_rdma_dev == 0)
+        {
+            return { 0 };
+        }
+        return { raw->rdma_dev_info[0].num_rdma_ports };
+    }
+
+    /**
+     * @brief Get NIC RDMA port statistics.
+     * @param rdma_port_idx RDMA port index.
+     * @return Vector of stat entries (empty if no statistics available).
+     * @throws std::runtime_error If AMD SMI query fails.
+     */
+    [[nodiscard]] std::vector<nic::stat_entry> get_nic_rdma_port_statistics(
+        std::uint8_t rdma_port_idx) const
+    {
+        std::uint32_t num_stats = 0;
+        check(amdsmi_get_nic_rdma_port_statistics(m_handle, rdma_port_idx, &num_stats,
+                                                  nullptr),
+              "get_nic_rdma_port_statistics (count)");
+
+        if(num_stats == 0)
+        {
+            return {};
+        }
+
+        std::vector<amdsmi_nic_stat_t> raw(num_stats);
+        check(amdsmi_get_nic_rdma_port_statistics(m_handle, rdma_port_idx, &num_stats,
+                                                  raw.data()),
+              "get_nic_rdma_port_statistics (data)");
+
+        std::vector<nic::stat_entry> result;
+        result.reserve(num_stats);
+        for(const auto& stat : raw)
+        {
+            result.emplace_back(nic::stat_entry{ stat.name, stat.value });
+        }
+        return result;
+    }
+#endif
+
 private:
-    static void check(amdsmi_status_t status, const char* func)
+    static void check(amdsmi_status_t status, std::string_view func)
     {
         if(status != AMDSMI_STATUS_SUCCESS)
         {
