+++ projects/rocprofiler-systems/source/lib/rocprof-sys/library/pmc/collectors/gpu/tests/mock_gpu_backend.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "backends/amd_smi/gpu_backend.hpp"
+#include "backends/amd_smi/device_backend.hpp"
 
 #include <gmock/gmock.h>
